requirements.txt
# Core runtime dependencies (install PyTorch separately; see README.md)

numpy>=1.24
python-chess>=1.999

tqdm>=4.66
pandas
